requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra dev -o requirements-dev.txt
alembic==1.13.1
    # via fast-log-queue (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.3.0
    # via
    #   fast-log-queue (pyproject.toml)
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via fast-log-queue (pyproject.toml)
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-lib==2.143.0
    # via fast-log-queue (pyproject.toml)
boto3==1.34.113
    # via
    #   fast-log-queue (pyproject.toml)
    #   moto
botocore==1.34.113
    # via
    #   boto3
    #   moto
    #   s3transfer
cattrs==23.2.3
    # via jsii
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
constructs==10.3.0
    # via aws-cdk-lib
cryptography==42.0.7
    # via moto
distlib==0.3.8
    # via virtualenv
dnspython==2.6.1
    # via email-validator
docker==7.1.0
    # via testcontainers
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via fast-log-queue (pyproject.toml)
fastapi-cli==0.0.4
    # via fastapi
filelock==3.14.0
    # via virtualenv
greenlet==3.0.3
    # via fast-log-queue (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fast-log-queue (pyproject.toml)
    #   fastapi
identify==2.5.36
    # via pre-commit
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-resources==6.4.0
    # via jsii
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via
    #   fastapi
    #   moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.98.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   constructs
mako==1.3.5
    # via alembic
mangum==0.17.0
    # via fast-log-queue (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
moto==5.0.8
    # via fast-log-queue (pyproject.toml)
nodeenv==1.9.0
    # via pre-commit
orjson==3.10.3
    # via fastapi
packaging==24.0
    # via pytest
platformdirs==4.2.2
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.7.1
    # via fast-log-queue (pyproject.toml)
psycopg2-binary==2.9.9
    # via fast-log-queue (pyproject.toml)
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   constructs
    #   jsii
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-extra-types==2.7.0
    # via fastapi
pydantic-settings==2.2.1
    # via fastapi
pygments==2.18.0
    # via rich
pytest==8.2.1
    # via fast-log-queue (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
    #   moto
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   fastapi
    #   pre-commit
    #   responses
    #   uvicorn
requests==2.32.2
    # via
    #   docker
    #   moto
    #   responses
responses==0.25.0
    # via moto
rich==13.7.1
    # via typer
s3transfer==0.10.1
    # via boto3
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.30
    # via
    #   fast-log-queue (pyproject.toml)
    #   alembic
starlette==0.37.2
    # via fastapi
testcontainers==4.5.0
    # via fast-log-queue (pyproject.toml)
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   constructs
    #   jsii
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.0
    # via
    #   alembic
    #   fastapi
    #   jsii
    #   mangum
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   testcontainers
    #   typer
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via
    #   botocore
    #   docker
    #   requests
    #   responses
    #   testcontainers
uvicorn==0.29.0
    # via fastapi
uvloop==0.19.0
    # via uvicorn
virtualenv==20.26.2
    # via pre-commit
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
werkzeug==3.0.3
    # via moto
wrapt==1.16.0
    # via testcontainers
xmltodict==0.13.0
    # via moto
